Weather/Rasp: Refactor color tables into separate file To prepare for functional code to move into RaspStyle.cpp
